+++ src/Component/OpenApi2/Tests/fixtures/whitelisted-paths-array-notation/expected/Runtime/Client/Client.php
@@ -44,6 +44,9 @@ abstract class Client
      */
     public function stream(iterable|ResponseInterface $responses, ?float $timeout = null): ResponseStreamInterface
     {
+        if ($responses instanceof ResponseInterface) {
+            $responses = [$responses];
+        }
         $mapped = [];
         foreach ($responses as $response) {
             $mapped[] = $response instanceof Result ? $response->getResponse() : $response;