+++ src/Component/OpenApi3/Tests/fixtures/fetch-mode-default/expected/Client.php
@@ -30,6 +30,10 @@ class Client extends \Jane\Component\OpenApi3\Tests\FetchModeDefault\Runtime\Cli
     {
         return $this->executeEndpoint(new \Jane\Component\OpenApi3\Tests\FetchModeDefault\Endpoint\GetPet($petId));
     }
+    /**
+     * @param list<callable(\Symfony\Contracts\HttpClient\HttpClientInterface): \Symfony\Contracts\HttpClient\HttpClientInterface> $additionalPlugins HttpClientInterface decorator factories, applied left-to-right after the server URL decorator
+     * @param list<\Symfony\Component\Serializer\Normalizer\NormalizerInterface|\Symfony\Component\Serializer\Normalizer\DenormalizerInterface> $additionalNormalizers
+     */
     public static function create(?\Symfony\Contracts\HttpClient\HttpClientInterface $httpClient = null, array $additionalPlugins = [], array $additionalNormalizers = [])
     {
         if (null === $httpClient) {