+++ src/Component/OpenApi3/Tests/fixtures/issue-810/expected/Client.php
@@ -14,6 +14,10 @@ class Client extends \Jane\Component\OpenApi3\Tests\Expected\Issue810\Runtime\Cl
     {
         return $this->executeEndpoint(new \Jane\Component\OpenApi3\Tests\Expected\Issue810\Endpoint\MimeTypeGeneratedValidDocBlock($requestBody, $accept));
     }
+    /**
+     * @param list<callable(\Symfony\Contracts\HttpClient\HttpClientInterface): \Symfony\Contracts\HttpClient\HttpClientInterface> $additionalPlugins HttpClientInterface decorator factories, applied left-to-right after the server URL decorator
+     * @param list<\Symfony\Component\Serializer\Normalizer\NormalizerInterface|\Symfony\Component\Serializer\Normalizer\DenormalizerInterface> $additionalNormalizers
+     */
     public static function create(?\Symfony\Contracts\HttpClient\HttpClientInterface $httpClient = null, array $additionalPlugins = [], array $additionalNormalizers = [])
     {
         if (null === $httpClient) {