+++ src/Component/OpenApi3/Tests/fixtures/whitelisted-paths-array-notation/expected/Client.php
@@ -36,6 +36,10 @@ class Client extends \Jane\OpenApi3\Tests\Expected\WhitelistedPathsArrayNotation
     {
         return $this->executeEndpoint(new \Jane\OpenApi3\Tests\Expected\WhitelistedPathsArrayNotation\Endpoint\AddOrDeleteRules($requestBody, $queryParameters, $accept));
     }
+    /**
+     * @param list<callable(\Symfony\Contracts\HttpClient\HttpClientInterface): \Symfony\Contracts\HttpClient\HttpClientInterface> $additionalPlugins HttpClientInterface decorator factories, applied left-to-right after the server URL decorator
+     * @param list<\Symfony\Component\Serializer\Normalizer\NormalizerInterface|\Symfony\Component\Serializer\Normalizer\DenormalizerInterface> $additionalNormalizers
+     */
     public static function create(?\Symfony\Contracts\HttpClient\HttpClientInterface $httpClient = null, array $additionalPlugins = [], array $additionalNormalizers = [], bool $applyServerPlugins = true)
     {
         $plugins = [];