+++ src/Component/OpenApi31/Tests/fixtures/issue-1036/expected/Client.php
@@ -13,6 +13,10 @@ class Client extends \Jane\Component\OpenApi31\Tests\ExpectedIssue1036\Runtime\C
     {
         return $this->executeEndpoint(new \Jane\Component\OpenApi31\Tests\ExpectedIssue1036\Endpoint\UploadDocument($requestBody));
     }
+    /**
+     * @param list<callable(\Symfony\Contracts\HttpClient\HttpClientInterface): \Symfony\Contracts\HttpClient\HttpClientInterface> $additionalPlugins HttpClientInterface decorator factories, applied left-to-right after the server URL decorator
+     * @param list<\Symfony\Component\Serializer\Normalizer\NormalizerInterface|\Symfony\Component\Serializer\Normalizer\DenormalizerInterface> $additionalNormalizers
+     */
     public static function create(?\Symfony\Contracts\HttpClient\HttpClientInterface $httpClient = null, array $additionalPlugins = [], array $additionalNormalizers = [])
     {
         if (null === $httpClient) {