+++ src/Component/OpenApi31/Tests/fixtures/issue-939/expected/Client.php
@@ -12,6 +12,10 @@ class Client extends \Jane\Component\OpenApi31\Tests\Expected\Issue939\Runtime\C
     {
         return $this->executeEndpoint(new \Jane\Component\OpenApi31\Tests\Expected\Issue939\Endpoint\GetProposal());
     }
+    /**
+     * @param list<callable(\Symfony\Contracts\HttpClient\HttpClientInterface): \Symfony\Contracts\HttpClient\HttpClientInterface> $additionalPlugins HttpClientInterface decorator factories, applied left-to-right after the server URL decorator
+     * @param list<\Symfony\Component\Serializer\Normalizer\NormalizerInterface|\Symfony\Component\Serializer\Normalizer\DenormalizerInterface> $additionalNormalizers
+     */
     public static function create(?\Symfony\Contracts\HttpClient\HttpClientInterface $httpClient = null, array $additionalPlugins = [], array $additionalNormalizers = [], bool $applyServerPlugins = true)
     {
         $plugins = [];