+++ src/Component/OpenApiCommon/Generator/Runtime/Client/FormEncoder.php
@@ -23,6 +23,7 @@ final class FormEncoder implements EncoderInterface, DecoderInterface
 
     public function decode(string $data, string $format, array $context = []): mixed
     {
+        $result = [];
         \parse_str($data, $result);
 
         return $result;